# openapi.yaml
openapi: 3.1.0
info:
  title: Telescope MAX API
  version: 1.0.0
  description: >-
    REST API for the Telescope mini-app. Protected methods use MAX initData in
    X-Max-Init-Data.
servers:
  - url: https://max-bot-sable.vercel.app
    description: Deployed Vercel origin
security:
  - MaxInitData: []
paths:
  /api/auth/validate:
    post:
      security: []
      summary: Validate MAX mini-app initData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData]
              properties:
                initData: { type: string }
      responses:
        '200':
          description: Valid initData
          content:
            application/json:
              schema:
                type: object
                required: [ok, user]
                properties:
                  ok: { type: boolean, const: true }
                  user:
                    type: object
                    required: [id, name]
                    properties:
                      id: { type: integer, format: int64 }
                      name: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/me:
    get:
      summary: Get the current user profile and sharing status
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/consent:
    post:
      summary: Save consent status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/me/reminders:
    post:
      summary: Enable or disable reminders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/tasks:
    get:
      summary: List the current user's tasks
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/TaskStatus' }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, deadline]
              properties:
                title: { type: string, maxLength: 200 }
                subject: { type: string }
                deadline: { type: string, format: date-time }
                urgent: { type: boolean }
                important: { type: boolean }
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/tasks/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    patch:
      summary: Update task fields or status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                title: { type: string, maxLength: 200 }
                subject: { type: string }
                deadline: { type: string, format: date-time }
                urgent: { type: boolean }
                important: { type: boolean }
                status: { $ref: '#/components/schemas/TaskStatus' }
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: Delete a task
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/moods:
    get:
      summary: List mood entries for a number of days
      parameters:
        - in: query
          name: days
          schema: { type: integer, minimum: 1, default: 7 }
      responses:
        '200':
          description: Mood entries, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Mood' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Save a mood entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mood]
              properties:
                mood: { $ref: '#/components/schemas/MoodValue' }
                note: { type: string }
      responses:
        '201':
          description: Created mood entry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Mood' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/notes:
    get:
      summary: List the current user's notes
      responses:
        '200':
          description: Note list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Note' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, allDay, start, end]
              properties:
                text: { type: string }
                allDay: { type: boolean }
                start: { type: string, format: date-time }
                end: { type: string, format: date-time }
      responses:
        '201':
          description: Created note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/notes/{id}:
    delete:
      summary: Delete a note
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/group/members:
    get:
      summary: List participants who opted into group sharing
      responses:
        '200':
          description: Participants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/group/tasks:
    get:
      summary: List tasks shared by opted-in participants
      responses:
        '200':
          description: Shared tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/group/sharing:
    post:
      summary: Opt in or out of sharing tasks with the cohort
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          description: Updated sharing choice
          content:
            application/json:
              schema:
                type: object
                required: [enabled]
                properties:
                  enabled: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/tasks/scan:
    post:
      summary: Prepare a task draft from a photo (mock, no OCR)
      description: The current handler does not inspect the image; the title is derived from the uploaded filename.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo: { type: string, format: binary }
      responses:
        '200':
          description: Draft task data
          content:
            application/json:
              schema:
                type: object
                required: [title, subject, deadline]
                properties:
                  title: { type: string }
                  subject: { type: string }
                  deadline: { type: string, format: date-time }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/feedback:
    post:
      summary: Submit feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/webhook:
    post:
      security: []
      summary: Receive MAX bot updates
      responses:
        '200': { description: Update accepted }
  /api/cron:
    get:
      security: []
      summary: Dispatch due reminders (Vercel Cron)
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: { type: string, description: Bearer CRON_SECRET }
      responses:
        '200':
          description: Dispatch result
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: Cron is not configured }
components:
  securitySchemes:
    MaxInitData:
      type: apiKey
      in: header
      name: X-Max-Init-Data
      description: MAX WebApp initData. Not required for auth validation, webhook, or cron.
  schemas:
    TaskStatus:
      type: string
      enum: [new, in_progress, done]
    MoodValue:
      type: string
      enum: [good, ok, bad]
    Task:
      type: object
      required: [id, user_id, title, subject, deadline, urgent, important, status]
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        title: { type: string }
        subject: { type: string }
        deadline: { type: string, format: date-time }
        urgent: { type: boolean }
        important: { type: boolean }
        status: { $ref: '#/components/schemas/TaskStatus' }
        group: { type: boolean }
    Mood:
      type: object
      required: [id, user_id, value, created_at]
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        value: { $ref: '#/components/schemas/MoodValue' }
        note: { type: string }
        created_at: { type: string, format: date-time }
    Note:
      type: object
      required: [id, user_id, text, allDay, start, end]
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        text: { type: string }
        allDay: { type: boolean }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
    Profile:
      type: object
      required: [id, name, consent, onboarded, reminders_on, group_sharing]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        consent: { type: boolean }
        onboarded: { type: boolean }
        reminders_on: { type: boolean }
        group_sharing: { type: boolean }
    Member:
      type: object
      required: [id, name, group, tasks]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        group: { type: string }
        tasks: { type: integer }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Ok:
      description: Operation succeeded
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid MAX initData
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Invalid task status transition
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }